Improve feedback for user. Improve UI and fix other bugs
Change output of listbox to not spam as much and have more helpful
information.

Change map marker so the title fits on the screen.

Make the UI better.

Remove the interface that handles sending data from the mobile to
arduino because it's not needed.

Make it so data only uploads to server if there is new data.

Give feedback about the file uploading.